+++ rollup.config.mjs
@@ -2,6 +2,7 @@ import replace from "@rollup/plugin-replace";
 import bundleSize from "rollup-plugin-bundle-size";
 import commonjs from "@rollup/plugin-commonjs";
 import resolve from "@rollup/plugin-node-resolve";
+import json from "@rollup/plugin-json";
 import pkg from "./package.json" assert { type: "json" };
 
 // credit/license information
@@ -18,6 +19,7 @@ const license = [
 export default [{
     input: "src/index.js",
     plugins: [
+        json(),
         resolve({
             mainFields: ["module"],
             browser: true,